Add files via upload
The entirety of the website, along with its construction, is articulated in my native tongue, Bulgarian. Its focal purpose is to facilitate donations for individuals experiencing homelessness and those in various states of need. Any ideas for updates about this website's realization will be thankful cause this website may be the change for many people for something better in their lifes!